+++ gift_certificates/src/main/java/com/epam/esm/giftCertificate/GiftCertificateRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GiftCertificateRepository extends JpaRepository<GiftCertificate, Integer> {
+public interface GiftCertificateRepository extends JpaRepository<GiftCertificate, Long> {
 }
